adjustments in category in-memo repository
